noenv_requirements.txt
--trusted-host dist.pinaxproject.com
--extra-index-url=http://dist.pinaxproject.com/dev/
--extra-index-url=http://dist.pinaxproject.com/alpha/

South
django_debug_toolbar
django-extensions
pinax-theme-bootstrap==2.0.4
pinax-theme-bootstrap-account==1.0b2
django-user-accounts==1.0b3
django-forms-bootstrap==2.0.3.post1
metron==1.0
pinax-utils==1.0b1.dev3
webtest
django-webtest
django-notification
toposort
django-cors-headers